Correct handling of multiline installing service
I was instaling OrientDB as service like in this doc and encountered issues. New lines with `\` wasn't working for me, service was installed, but without all parameters after `--DisplayName`. Additionally error message was throw to the console: `'--Description'is not recognized as an internal or external command, operable program or batch file.`

As far as I know caret (`^`) should be used for concatenating lines.