UTF-8 charset as default in Windows service
As stated in Web Server docs, OrientDB server should use `UTF-8` charset as default. But on Windows with service created like in this docs, it is not correct.

When I save `ąćęłńóśżź` (polish letters) into my vertex, I get `Ä…Ä‡Ä™Ĺ‚Ĺ„ĂłĹ›ĹĽĹş` as a result.

When I stop the service and run server with `bin/server.bat` and save the same value, it is stored correctly.

I think it's because `installService.bat` misses `-Dfile.Encoding=UTF8` (compared to `server.bat`).