Update SQL-Create-Function.md
Using square brackets for the list is very confusing. It should probably be parenthesis because that is what you use when creating an index and need to specify a list of properties. However since that will probably never change because of backwards compatibility.

You need to specify that the square brackets are required because usually the square brackets indicate something optional which is where the current documentation is miss leading. I added a sentence to the syntax PARAMETERS description to help with this as well as added an example to help clarify.