Edits to Tutorial Doc
I added some graphics, and made minor corrections (inserted missing
apostrophes etc. where appropriate). Removed the sand-alone wikipedia
link and integrated it into the tutorial text.

Notice on classes.md I removed references to fields and properties.. I
could not find them mentioned anywhere else in the documentations. I
couldn’t make sense of whether a field was a key and a value was a
proper.

Maybe this is an important topic that should stay in the documentation
though. Can someone else explain this concept?