updated clusters tutorial and attempt to make code in *pre* wrappers stop breaking in line
I want to take out references to older versions of orientDB. I think
that if you are running through the tutorial it's okay to assume that
you're running the latest version.

Updated references to the number of clusters a class is created with to
reflect that it depends on the number of CPU cores.

Added introduction and conclusion to the page to give the tutorial a
continuous story-like flow.